separated metadata viewer component from image tool  for better  code quality
