+++ package.json
@@ -26,6 +26,7 @@
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "dayjs": "^1.11.13",
+    "exifr": "^7.1.3",
     "jspdf": "^3.0.1",
     "jspdf-autotable": "^5.0.2",
     "lucide-react": "^0.534.0",
@@ -39,6 +40,7 @@
     "react-day-picker": "^9.8.1",
     "react-dom": "^19.1.1",
     "react-dropzone": "^14.3.8",
+    "react-easy-crop": "^5.5.0",
     "react-hook-form": "^7.61.1",
     "react-icons": "^5.5.0",
     "react-markdown": "^10.1.0",
